- added a 'new groups' block. this block and the subscribers block are now configurable as to the max number of items that they display. - added 'my groups' block
each group may now decide if it wants to be listed on the 'list of groups' page and block. this commit added the 'directory' column to the og table. you will need to re-save your groups or run a SQL statement to set this field to 0 if you want hidden groups. a byproduct of this change is that all group home pages are now accessible (as they once were). The content on those pages is shown/hidden according to your 'visibility' settings. once again, you might want to re-save existing groups which were INVITE ONLY. that will make their home pages accessible to non-subscribers.

changed 'register' field to a checkbox on aadmin/settings/og page

fix: groups which don't choose otherwise follow any change in site default theme